Handle plural strings from CDS
During pushing, when we capture a simple pluralized ICU string, the CDS
will push it to transifex like a pluralized string:

    js app -> cds: {cnt, plural, one {ONE} other {OTHER}}

    cds -> APIv3:  {..., strings: {one: "ONE", other: "OTHER"}, ...}

During pulling, the CDS will format the pluralized string back to an ICU
pluralized string. Unfortunately, the variable name has been lost by
this point, so CDS will replace it with '???'

    cds -> js app's cache: {???, plural, one {ΕΝΑ} other {ΑΛΛΟ}}

When rendering a translation, we expect the source string's key to match
the one in the cache and we can extract the variable name to the source
string. So before we feed the translation to the MessageFormat library,
we replace '???' with the variable name:

    t('{cnt, plural, one {ONE} other {OTHER}}', { cnt: 8 })

    key = generateKey(sourceString)
    translation = cache.get(key)  // {???, plural, one {ΕΝΑ} other {ΑΛΛΟ}}
    if (isPluralized(translation)) {
      translation = `{${varName}${translation.substring(4)}`
      // {cnt, plural, one {ΕΝΑ} other {ΑΛΛΟ}}
    }
    msg = MF.compile(translation)
    return msg(params)