Fix interpolation in psuedolocalization policy
This commit resolves an issue in pseudolocalization where interpolation
of framework components would be broken. This does introduce the concept
of interpolation into the native framework where it wasn't before, which
we consider acceptable given that interpolation will be needed for
several different framework libraries, so it's better to not have to
duplicate this in those.