'useT' hook to return function
Note: This is a breaking change and should be communicated as such!!!
Fortunately, I doubt many users are using 'useT' at the moment so we
should be ok.

The previous 'useT' implementation returned the translated string. The
problem was that, since it's a hook, it must be used at the top of a
component before any control flows. So this would be dangerous:

```javascript
function AComponent({ ... }) {
  if (something) {
    doSomething();
  } else {
    doSomethingElse();
  }
  const message = useT(...);
  return <b>{message}<b>
}
```

since if any hooks are used in the if/else block, it would affect the
order with which the hooks are used. In fact, I suspect that eslint can
raise an error if the above snippet is used.

This implementation:

1. Forces the current component to rerender when the language changes
2. Returns a reference to 'translateWithElements'

Side note: this is more in line with what react-i18next offers.

The correct way of using useT after this is that instead of:

```javascript
function MyComponent() {
  const message = useT('Hello world');
  return message;
}
```

You should write this

```javascript
function MyComponent() {
  const t = useT();
  const message = t('Hello world');
  return message;
}
```