Add Vue parsing to CLI & refactor extract function
In order to parse for phrases the Vue SFC we needed to create a new
extract function. This means that cli will also be parsing all the
`*.vue` files that will find.

`vue-template-compiler` was used to do this so we can extract the
different types of content from the Vue template `script` and
`template`.

For the `template` part we use `compile` to get the AST and we traverse
with a custom function and for the `script` we just send to babel.

During this implementation it seemed that a lot of code could be
extracted from `extract.js` and so a new `parsers` folder was created
with `angularHTML`, `vue` and `babel` files with all the functionality
that can be reused accross parsers if needed.