Add support for Vue3 plugin & update CLI
In order to support Vue3 we had to drop `vue-template-compiler`
in favor of `@vue/compiler-sfc` for compatibility reasons. This
way we can now support both Vue2 and Vue3.

Side effects
`vue/compiler-sfc` ast result is more cryptic than the previous
dependency, so we might need to revisit in the future and
use an alternative.