Revert change about additional type (Vue)
In a previous commit we included in the Vue parser
the type `12` in visitors in order to parse some
simblings and pass the content to the BabelParser.

This resulted to simple text being parsed and Babel
throwing an error when trying to parse for strings.

In this fix in case there are no `children` we use
`traverseNode` to the `content` which should result
to parsing all the things needed and causing no errors.